+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Change `Row::Base` to an abstract class
 * Add support Ruby v2.5
+* Add file options to `Reader::Core#each`
 
 ### Misc